refactor: 抽离独立use文件,抽离独立props类型文件
